chore: refactor Browser.js into seperate files
