+++ lib/Launcher.js
@@ -19,7 +19,7 @@ const removeFolder = require('rimraf');
 const childProcess = require('child_process');
 const BrowserFetcher = require('./BrowserFetcher');
 const {Connection} = require('./Connection');
-const {Browser} = require('./Browser');
+const Browser = require('./Browser');
 const readline = require('readline');
 const fs = require('fs');
 const {helper} = require('./helper');
