+++ lib/externs.d.ts
@@ -1,6 +1,8 @@
 import { Connection as RealConnection, CDPSession as RealCDPSession } from './Connection.js';
-import {Browser as RealBrowser, TaskQueue as RealTaskQueue, Target as RealTarget} from './Browser.js';
+import * as RealBrowser from './Browser.js';
+import * as RealTarget from './Target.js';
 import * as RealPage from './Page.js';
+import * as RealTaskQueue from './TaskQueue.js';
 import {Mouse as RealMouse, Keyboard as RealKeyboard, Touchscreen as RealTouchscreen}  from './Input.js';
 import {Frame as RealFrame, FrameManager as RealFrameManager}  from './FrameManager.js';
 import {JSHandle as RealJSHandle, ExecutionContext as RealExecutionContext}  from './ExecutionContext.js';
@@ -28,6 +30,14 @@ export class Page extends RealPage {}
 export interface ConnectionTransport extends NodeJS.EventEmitter {
   send(string);
   close();
+}
+
+export interface TargetInfo {
+  type: string;
+  targetId: string;
+  title: string;
+  url: string;
+  attached: boolean;
 }
 
 export interface ChildProcess extends child_process.ChildProcess {}
